requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.25.9
python-dotenv==1.0.0
schedule==1.2.0
python-dateutil==2.8.2